Update build process
