+++ tasks/tscc/tsconfig.json
@@ -1,11 +1,11 @@
 {
     "compilerOptions": {
-        "target": "es5",
+        "target": "es6",
         "lib": ["DOM", "ES6"],
-        "outDir": "../../build/tsc",
+        "outDir": "../../build/tsickle",
         "module": "commonjs"
     },
     "include": [
-        "../../build/tsickle/**/*.ts"
+        "../../build/prep/**/*.ts"
     ]
 }
