+++ test/dom/HTMLIFrame.ts
@@ -8,15 +8,15 @@ import { TLEventType } from '../../src/timeline/event';
 const expect = chai.expect;
 
 describe('HTMLIFrameElement#src', function() {
-    it('should log when an iframe\'s src attribute changes', function() {
+    it('should log when an iframe\'s contentDocument property is accessed', function() {
         let ifr = document.createElement('iframe');
         ifr.style.cssText = 'display:none;';
         document.body.appendChild(ifr);
-        ifr.src = 'about:blank';
+        if (!('location' in ifr.contentDocument)) { throw new Error(); }
         let records = timeline.takeRecords()[0];
         console.log(records);
         let lastEvent = records[records.length - 1];
         document.body.removeChild(ifr);
-        expect(lastEvent.type).to.equal(TLEventType.SET);
+        expect(lastEvent.$type).to.equal(TLEventType.GET);
     });
 });
