+++ test/timeline/index.ts
@@ -9,6 +9,6 @@ describe('Timeline', function() {
     it('should log when it is first created', function() {
         let records = timeline.takeRecords();
         let firstEvent = timeline.takeRecords()[0][0];
-        expect(firstEvent.type).to.equal(TLEventType.CREATE);
+        expect(firstEvent.$type).to.equal(TLEventType.CREATE);
     });
 });
